--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "tree-sitter-elm"]
-	path = tree-sitter-elm
-	url = https://github.com/Razzeee/tree-sitter-elm
