+++ Cargo.toml
@@ -19,6 +19,7 @@ thiserror = "1.0.21"
 tree-sitter = "0.17.0"
 regex = "1.4.1"
 lazy_static = "1.4.0"
+tree-sitter-elm = "4.3.4"
 
 [dev-dependencies]
