+++ src/run.rs
@@ -179,7 +179,7 @@ pub fn main(options: Options) {
 
     // Find all modules and tests
     eprintln!("Finding all modules and tests ...");
-    let all_modules_and_tests = crate::elmi::all_tests(&module_paths).unwrap();
+    let all_modules_and_tests = crate::parser::all_tests(&module_paths).unwrap();
 
     let (runner_imports, maybe_runner_tests): (Vec<String>, Vec<String>) = all_modules_and_tests
         .iter()
